[FLINK-11730] [State Backends] Make HeapKeyedStateBackend follow the builder pattern
This closes #7866.

(cherry picked from commit 237d07c76b51c171f0f41f9f82d777df26da1dd4)
