+++ flink-runtime/src/main/java/org/apache/flink/runtime/state/heap/HeapKeyedStateBackend.java
@@ -32,64 +32,39 @@ import org.apache.flink.api.common.typeutils.TypeSerializer;
 import org.apache.flink.api.common.typeutils.TypeSerializerSchemaCompatibility;
 import org.apache.flink.api.java.tuple.Tuple2;
 import org.apache.flink.core.fs.CloseableRegistry;
-import org.apache.flink.core.fs.FSDataInputStream;
-import org.apache.flink.core.memory.DataInputViewStreamWrapper;
-import org.apache.flink.core.memory.DataOutputViewStreamWrapper;
 import org.apache.flink.runtime.checkpoint.CheckpointOptions;
 import org.apache.flink.runtime.query.TaskKvStateRegistry;
 import org.apache.flink.runtime.state.AbstractKeyedStateBackend;
-import org.apache.flink.runtime.state.AbstractSnapshotStrategy;
-import org.apache.flink.runtime.state.AsyncSnapshotCallable;
 import org.apache.flink.runtime.state.CheckpointStreamFactory;
-import org.apache.flink.runtime.state.CheckpointStreamWithResultProvider;
-import org.apache.flink.runtime.state.CheckpointedStateScope;
-import org.apache.flink.runtime.state.DoneFuture;
 import org.apache.flink.runtime.state.KeyExtractorFunction;
 import org.apache.flink.runtime.state.KeyGroupRange;
-import org.apache.flink.runtime.state.KeyGroupRangeOffsets;
 import org.apache.flink.runtime.state.KeyGroupedInternalPriorityQueue;
-import org.apache.flink.runtime.state.KeyGroupsStateHandle;
 import org.apache.flink.runtime.state.Keyed;
-import org.apache.flink.runtime.state.KeyedBackendSerializationProxy;
 import org.apache.flink.runtime.state.KeyedStateFunction;
 import org.apache.flink.runtime.state.KeyedStateHandle;
 import org.apache.flink.runtime.state.LocalRecoveryConfig;
 import org.apache.flink.runtime.state.PriorityComparable;
 import org.apache.flink.runtime.state.RegisteredKeyValueStateBackendMetaInfo;
 import org.apache.flink.runtime.state.RegisteredPriorityQueueStateBackendMetaInfo;
-import org.apache.flink.runtime.state.SnappyStreamCompressionDecorator;
 import org.apache.flink.runtime.state.SnapshotResult;
-import org.apache.flink.runtime.state.StateSnapshot;
-import org.apache.flink.runtime.state.StateSnapshotKeyGroupReader;
+import org.apache.flink.runtime.state.StateSerializerProvider;
 import org.apache.flink.runtime.state.StateSnapshotRestore;
 import org.apache.flink.runtime.state.StateSnapshotTransformer.StateSnapshotTransformFactory;
 import org.apache.flink.runtime.state.StateSnapshotTransformers;
 import org.apache.flink.runtime.state.StreamCompressionDecorator;
-import org.apache.flink.runtime.state.StreamStateHandle;
-import org.apache.flink.runtime.state.UncompressedStreamCompressionDecorator;
-import org.apache.flink.runtime.state.metainfo.StateMetaInfoSnapshot;
 import org.apache.flink.runtime.state.ttl.TtlTimeProvider;
 import org.apache.flink.util.FlinkRuntimeException;
-import org.apache.flink.util.Preconditions;
 import org.apache.flink.util.StateMigrationException;
-import org.apache.flink.util.function.SupplierWithException;
 
-import org.apache.commons.io.IOUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nonnull;
 
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.util.ArrayList;
 import java.util.Collection;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
-import java.util.concurrent.FutureTask;
 import java.util.concurrent.RunnableFuture;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -132,7 +107,7 @@ public class HeapKeyedStateBackend<K> extends AbstractKeyedStateBackend<K> {
 	/**
 	 * The snapshot strategy for this backend. This determines, e.g., if snapshots are synchronous or asynchronous.
 	 */
-	private final HeapSnapshotStrategy snapshotStrategy;
+	private final HeapSnapshotStrategy<K> snapshotStrategy;
 
 	/**
 	 * Factory for state that is organized as priority queue.
@@ -141,31 +116,28 @@ public class HeapKeyedStateBackend<K> extends AbstractKeyedStateBackend<K> {
 
 	public HeapKeyedStateBackend(
 		TaskKvStateRegistry kvStateRegistry,
-		TypeSerializer<K> keySerializer,
+		StateSerializerProvider<K> keySerializerProvider,
 		ClassLoader userCodeClassLoader,
 		int numberOfKeyGroups,
 		KeyGroupRange keyGroupRange,
-		boolean asynchronousSnapshots,
 		ExecutionConfig executionConfig,
+		TtlTimeProvider ttlTimeProvider,
+		CloseableRegistry cancelStreamRegistry,
+		StreamCompressionDecorator keyGroupCompressionDecorator,
+		Map<String, StateTable<K, ?, ?>> registeredKVStates,
+		Map<String, HeapPriorityQueueSnapshotRestoreWrapper> registeredPQStates,
 		LocalRecoveryConfig localRecoveryConfig,
 		HeapPriorityQueueSetFactory priorityQueueSetFactory,
-		TtlTimeProvider ttlTimeProvider,
-		CloseableRegistry cancelStreamRegistry) {
-
-		super(kvStateRegistry, keySerializer, userCodeClassLoader,
-			numberOfKeyGroups, keyGroupRange, executionConfig, ttlTimeProvider, new CloseableRegistry());
-
-		this.registeredKVStates = new HashMap<>();
-		this.registeredPQStates = new HashMap<>();
-		this.localRecoveryConfig = Preconditions.checkNotNull(localRecoveryConfig);
-
-		SnapshotStrategySynchronicityBehavior<K> synchronicityTrait = asynchronousSnapshots ?
-			new AsyncSnapshotStrategySynchronicityBehavior() :
-			new SyncSnapshotStrategySynchronicityBehavior();
-
-		this.snapshotStrategy = new HeapSnapshotStrategy(synchronicityTrait);
+		HeapSnapshotStrategy<K> snapshotStrategy
+	) {
+		super(kvStateRegistry, keySerializerProvider, userCodeClassLoader, numberOfKeyGroups,
+			keyGroupRange, executionConfig, ttlTimeProvider, cancelStreamRegistry, keyGroupCompressionDecorator);
+		this.registeredKVStates = registeredKVStates;
+		this.registeredPQStates = registeredPQStates;
+		this.localRecoveryConfig = localRecoveryConfig;
 		LOG.info("Initializing heap keyed state backend with stream factory.");
 		this.priorityQueueSetFactory = priorityQueueSetFactory;
+		this.snapshotStrategy = snapshotStrategy;
 	}
 
 	// ------------------------------------------------------------------------
@@ -227,9 +199,9 @@ public class HeapKeyedStateBackend<K> extends AbstractKeyedStateBackend<K> {
 	}
 
 	private <N, V> StateTable<K, N, V> tryRegisterStateTable(
-			TypeSerializer<N> namespaceSerializer,
-			StateDescriptor<?, V> stateDesc,
-			@Nonnull StateSnapshotTransformFactory<V> snapshotTransformFactory) throws StateMigrationException {
+		TypeSerializer<N> namespaceSerializer,
+		StateDescriptor<?, V> stateDesc,
+		@Nonnull StateSnapshotTransformFactory<V> snapshotTransformFactory) throws StateMigrationException {
 
 		@SuppressWarnings("unchecked")
 		StateTable<K, N, V> stateTable = (StateTable<K, N, V>) registeredKVStates.get(stateDesc.getName());
@@ -265,7 +237,7 @@ public class HeapKeyedStateBackend<K> extends AbstractKeyedStateBackend<K> {
 				newStateSerializer,
 				snapshotTransformFactory);
 
-			stateTable = snapshotStrategy.newStateTable(newMetaInfo);
+			stateTable = snapshotStrategy.newStateTable(this, newMetaInfo);
 			registeredKVStates.put(stateDesc.getName(), stateTable);
 		}
 
@@ -282,10 +254,6 @@ public class HeapKeyedStateBackend<K> extends AbstractKeyedStateBackend<K> {
 		final StateSnapshotRestore stateSnapshotRestore = registeredKVStates.get(state);
 		StateTable<K, N, ?> table = (StateTable<K, N, ?>) stateSnapshotRestore;
 		return table.getKeys(namespace);
-	}
-
-	private boolean hasRegisteredState() {
-		return !(registeredKVStates.isEmpty() && registeredPQStates.isEmpty());
 	}
 
 	@Override
@@ -337,189 +305,8 @@ public class HeapKeyedStateBackend<K> extends AbstractKeyedStateBackend<K> {
 	}
 
 	@SuppressWarnings("deprecation")
-	public void restore(Collection<KeyedStateHandle> restoredState) throws Exception {
-		if (restoredState == null || restoredState.isEmpty()) {
-			return;
-		}
-
-		LOG.info("Initializing heap keyed state backend from snapshot.");
-
-		if (LOG.isDebugEnabled()) {
-			LOG.debug("Restoring snapshot from state handles: {}.", restoredState);
-		}
-
-		restorePartitionedState(restoredState);
-	}
-
-	@SuppressWarnings({"unchecked"})
-	private void restorePartitionedState(Collection<KeyedStateHandle> state) throws Exception {
-
-		final Map<Integer, StateMetaInfoSnapshot> kvStatesById = new HashMap<>();
-		registeredKVStates.clear();
-		registeredPQStates.clear();
-
-		boolean keySerializerRestored = false;
-
-		for (KeyedStateHandle keyedStateHandle : state) {
-
-			if (keyedStateHandle == null) {
-				continue;
-			}
-
-			if (!(keyedStateHandle instanceof KeyGroupsStateHandle)) {
-				throw new IllegalStateException("Unexpected state handle type, " +
-						"expected: " + KeyGroupsStateHandle.class +
-						", but found: " + keyedStateHandle.getClass());
-			}
-
-			KeyGroupsStateHandle keyGroupsStateHandle = (KeyGroupsStateHandle) keyedStateHandle;
-			FSDataInputStream fsDataInputStream = keyGroupsStateHandle.openInputStream();
-			cancelStreamRegistry.registerCloseable(fsDataInputStream);
-
-			try {
-				DataInputViewStreamWrapper inView = new DataInputViewStreamWrapper(fsDataInputStream);
-
-				KeyedBackendSerializationProxy<K> serializationProxy =
-						new KeyedBackendSerializationProxy<>(userCodeClassLoader);
-
-				serializationProxy.read(inView);
-
-				if (!keySerializerRestored) {
-					// check for key serializer compatibility; this also reconfigures the
-					// key serializer to be compatible, if it is required and is possible
-					TypeSerializerSchemaCompatibility<K> keySerializerSchemaCompat =
-						checkKeySerializerSchemaCompatibility(serializationProxy.getKeySerializerSnapshot());
-					if (keySerializerSchemaCompat.isCompatibleAfterMigration() || keySerializerSchemaCompat.isIncompatible()) {
-						throw new StateMigrationException("The new key serializer must be compatible.");
-					}
-
-					keySerializerRestored = true;
-				}
-
-				List<StateMetaInfoSnapshot> restoredMetaInfos =
-					serializationProxy.getStateMetaInfoSnapshots();
-
-				createOrCheckStateForMetaInfo(restoredMetaInfos, kvStatesById);
-
-				readStateHandleStateData(
-					fsDataInputStream,
-					inView,
-					keyGroupsStateHandle.getGroupRangeOffsets(),
-					kvStatesById, restoredMetaInfos.size(),
-					serializationProxy.getReadVersion(),
-					serializationProxy.isUsingKeyGroupCompression());
-			} finally {
-				if (cancelStreamRegistry.unregisterCloseable(fsDataInputStream)) {
-					IOUtils.closeQuietly(fsDataInputStream);
-				}
-			}
-		}
-	}
-
-	private void readStateHandleStateData(
-		FSDataInputStream fsDataInputStream,
-		DataInputViewStreamWrapper inView,
-		KeyGroupRangeOffsets keyGroupOffsets,
-		Map<Integer, StateMetaInfoSnapshot> kvStatesById,
-		int numStates,
-		int readVersion,
-		boolean isCompressed) throws IOException {
-
-		final StreamCompressionDecorator streamCompressionDecorator = isCompressed ?
-			SnappyStreamCompressionDecorator.INSTANCE : UncompressedStreamCompressionDecorator.INSTANCE;
-
-		for (Tuple2<Integer, Long> groupOffset : keyGroupOffsets) {
-			int keyGroupIndex = groupOffset.f0;
-			long offset = groupOffset.f1;
-
-			// Check that restored key groups all belong to the backend.
-			Preconditions.checkState(keyGroupRange.contains(keyGroupIndex), "The key group must belong to the backend.");
-
-			fsDataInputStream.seek(offset);
-
-			int writtenKeyGroupIndex = inView.readInt();
-			Preconditions.checkState(writtenKeyGroupIndex == keyGroupIndex,
-				"Unexpected key-group in restore.");
-
-			try (InputStream kgCompressionInStream =
-					 streamCompressionDecorator.decorateWithCompression(fsDataInputStream)) {
-
-				readKeyGroupStateData(
-					kgCompressionInStream,
-					kvStatesById,
-					keyGroupIndex,
-					numStates,
-					readVersion);
-			}
-		}
-	}
-
-	private void readKeyGroupStateData(
-		InputStream inputStream,
-		Map<Integer, StateMetaInfoSnapshot> kvStatesById,
-		int keyGroupIndex,
-		int numStates,
-		int readVersion) throws IOException {
-
-		DataInputViewStreamWrapper inView =
-			new DataInputViewStreamWrapper(inputStream);
-
-		for (int i = 0; i < numStates; i++) {
-
-			final int kvStateId = inView.readShort();
-			final StateMetaInfoSnapshot stateMetaInfoSnapshot = kvStatesById.get(kvStateId);
-			final StateSnapshotRestore registeredState;
-
-			switch (stateMetaInfoSnapshot.getBackendStateType()) {
-				case KEY_VALUE:
-					registeredState = registeredKVStates.get(stateMetaInfoSnapshot.getName());
-					break;
-				case PRIORITY_QUEUE:
-					registeredState = registeredPQStates.get(stateMetaInfoSnapshot.getName());
-					break;
-				default:
-					throw new IllegalStateException("Unexpected state type: " +
-						stateMetaInfoSnapshot.getBackendStateType() + ".");
-			}
-
-			StateSnapshotKeyGroupReader keyGroupReader = registeredState.keyGroupReader(readVersion);
-			keyGroupReader.readMappingsInKeyGroup(inView, keyGroupIndex);
-		}
-	}
-
-	private void createOrCheckStateForMetaInfo(
-		List<StateMetaInfoSnapshot> restoredMetaInfo,
-		Map<Integer, StateMetaInfoSnapshot> kvStatesById) {
-
-		for (StateMetaInfoSnapshot metaInfoSnapshot : restoredMetaInfo) {
-			final StateSnapshotRestore registeredState;
-
-			switch (metaInfoSnapshot.getBackendStateType()) {
-				case KEY_VALUE:
-					registeredState = registeredKVStates.get(metaInfoSnapshot.getName());
-					if (registeredState == null) {
-						RegisteredKeyValueStateBackendMetaInfo<?, ?> registeredKeyedBackendStateMetaInfo =
-							new RegisteredKeyValueStateBackendMetaInfo<>(metaInfoSnapshot);
-						registeredKVStates.put(
-							metaInfoSnapshot.getName(),
-							snapshotStrategy.newStateTable(registeredKeyedBackendStateMetaInfo));
-					}
-					break;
-				case PRIORITY_QUEUE:
-					registeredState = registeredPQStates.get(metaInfoSnapshot.getName());
-					if (registeredState == null) {
-						createInternal(new RegisteredPriorityQueueStateBackendMetaInfo<>(metaInfoSnapshot));
-					}
-					break;
-				default:
-					throw new IllegalStateException("Unexpected state type: " +
-						metaInfoSnapshot.getBackendStateType() + ".");
-			}
-
-			if (registeredState == null) {
-				kvStatesById.put(kvStatesById.size(), metaInfoSnapshot);
-			}
-		}
+	public void restore(Collection<KeyedStateHandle> restoredState) {
+		// all restore work done in builder and nothing to do here
 	}
 
 	@Override
@@ -529,10 +316,10 @@ public class HeapKeyedStateBackend<K> extends AbstractKeyedStateBackend<K> {
 
 	@Override
 	public <N, S extends State, T> void applyToAllKeys(
-			final N namespace,
-			final TypeSerializer<N> namespaceSerializer,
-			final StateDescriptor<S, T> stateDescriptor,
-			final KeyedStateFunction<K, S> function) throws Exception {
+		final N namespace,
+		final TypeSerializer<N> namespaceSerializer,
+		final StateDescriptor<S, T> stateDescriptor,
+		final KeyedStateFunction<K, S> function) throws Exception {
 
 		try (Stream<K> keyStream = getKeys(stateDescriptor.getName(), namespace)) {
 
@@ -541,9 +328,9 @@ public class HeapKeyedStateBackend<K> extends AbstractKeyedStateBackend<K> {
 			final List<K> keys = keyStream.collect(Collectors.toList());
 
 			final S state = getPartitionedState(
-					namespace,
-					namespaceSerializer,
-					stateDescriptor);
+				namespace,
+				namespaceSerializer,
+				stateDescriptor);
 
 			for (K key : keys) {
 				setCurrentKey(key);
@@ -593,233 +380,6 @@ public class HeapKeyedStateBackend<K> extends AbstractKeyedStateBackend<K> {
 		return localRecoveryConfig;
 	}
 
-	private interface SnapshotStrategySynchronicityBehavior<K> {
-
-		default void finalizeSnapshotBeforeReturnHook(Runnable runnable) {
-
-		}
-
-
-		boolean isAsynchronous();
-
-		<N, V> StateTable<K, N, V> newStateTable(RegisteredKeyValueStateBackendMetaInfo<N, V> newMetaInfo);
-	}
-
-	private class AsyncSnapshotStrategySynchronicityBehavior implements SnapshotStrategySynchronicityBehavior<K> {
-
-		@Override
-		public boolean isAsynchronous() {
-			return true;
-		}
-
-		@Override
-		public <N, V> StateTable<K, N, V> newStateTable(RegisteredKeyValueStateBackendMetaInfo<N, V> newMetaInfo) {
-			return new CopyOnWriteStateTable<>(HeapKeyedStateBackend.this, newMetaInfo);
-		}
-	}
-
-	private class SyncSnapshotStrategySynchronicityBehavior implements SnapshotStrategySynchronicityBehavior<K> {
-
-		@Override
-		public void finalizeSnapshotBeforeReturnHook(Runnable runnable) {
-			// this triggers a synchronous execution from the main checkpointing thread.
-			runnable.run();
-		}
-
-		@Override
-		public boolean isAsynchronous() {
-			return false;
-		}
-
-		@Override
-		public <N, V> StateTable<K, N, V> newStateTable(RegisteredKeyValueStateBackendMetaInfo<N, V> newMetaInfo) {
-			return new NestedMapsStateTable<>(HeapKeyedStateBackend.this, newMetaInfo);
-		}
-	}
-
-	/**
-	 * Base class for the snapshots of the heap backend that outlines the algorithm and offers some hooks to realize
-	 * the concrete strategies. Subclasses must be threadsafe.
-	 */
-	private class HeapSnapshotStrategy
-		extends AbstractSnapshotStrategy<KeyedStateHandle> implements SnapshotStrategySynchronicityBehavior<K> {
-
-		private final SnapshotStrategySynchronicityBehavior<K> snapshotStrategySynchronicityTrait;
-
-		HeapSnapshotStrategy(
-			SnapshotStrategySynchronicityBehavior<K> snapshotStrategySynchronicityTrait) {
-			super("Heap backend snapshot");
-			this.snapshotStrategySynchronicityTrait = snapshotStrategySynchronicityTrait;
-		}
-
-		@Nonnull
-		@Override
-		public RunnableFuture<SnapshotResult<KeyedStateHandle>> snapshot(
-			long checkpointId,
-			long timestamp,
-			@Nonnull CheckpointStreamFactory primaryStreamFactory,
-			@Nonnull CheckpointOptions checkpointOptions) throws IOException {
-
-			if (!hasRegisteredState()) {
-				return DoneFuture.of(SnapshotResult.empty());
-			}
-
-			int numStates = registeredKVStates.size() + registeredPQStates.size();
-
-			Preconditions.checkState(numStates <= Short.MAX_VALUE,
-				"Too many states: " + numStates +
-					". Currently at most " + Short.MAX_VALUE + " states are supported");
-
-			final List<StateMetaInfoSnapshot> metaInfoSnapshots = new ArrayList<>(numStates);
-			final Map<StateUID, Integer> stateNamesToId =
-				new HashMap<>(numStates);
-			final Map<StateUID, StateSnapshot> cowStateStableSnapshots =
-				new HashMap<>(numStates);
-
-			processSnapshotMetaInfoForAllStates(
-				metaInfoSnapshots,
-				cowStateStableSnapshots,
-				stateNamesToId,
-				registeredKVStates,
-				StateMetaInfoSnapshot.BackendStateType.KEY_VALUE);
-
-			processSnapshotMetaInfoForAllStates(
-				metaInfoSnapshots,
-				cowStateStableSnapshots,
-				stateNamesToId,
-				registeredPQStates,
-				StateMetaInfoSnapshot.BackendStateType.PRIORITY_QUEUE);
-
-			final KeyedBackendSerializationProxy<K> serializationProxy =
-				new KeyedBackendSerializationProxy<>(
-					// TODO: this code assumes that writing a serializer is threadsafe, we should support to
-					// get a serialized form already at state registration time in the future
-					getKeySerializer(),
-					metaInfoSnapshots,
-					!Objects.equals(UncompressedStreamCompressionDecorator.INSTANCE, keyGroupCompressionDecorator));
-
-			final SupplierWithException<CheckpointStreamWithResultProvider, Exception> checkpointStreamSupplier =
-
-				localRecoveryConfig.isLocalRecoveryEnabled() ?
-
-					() -> CheckpointStreamWithResultProvider.createDuplicatingStream(
-						checkpointId,
-						CheckpointedStateScope.EXCLUSIVE,
-						primaryStreamFactory,
-						localRecoveryConfig.getLocalStateDirectoryProvider()) :
-
-					() -> CheckpointStreamWithResultProvider.createSimpleStream(
-						CheckpointedStateScope.EXCLUSIVE,
-						primaryStreamFactory);
-
-			//--------------------------------------------------- this becomes the end of sync part
-
-			final AsyncSnapshotCallable<SnapshotResult<KeyedStateHandle>> asyncSnapshotCallable =
-				new AsyncSnapshotCallable<SnapshotResult<KeyedStateHandle>>() {
-					@Override
-					protected SnapshotResult<KeyedStateHandle> callInternal() throws Exception {
-
-						final CheckpointStreamWithResultProvider streamWithResultProvider =
-							checkpointStreamSupplier.get();
-
-						snapshotCloseableRegistry.registerCloseable(streamWithResultProvider);
-
-						final CheckpointStreamFactory.CheckpointStateOutputStream localStream =
-							streamWithResultProvider.getCheckpointOutputStream();
-
-						final DataOutputViewStreamWrapper outView = new DataOutputViewStreamWrapper(localStream);
-						serializationProxy.write(outView);
-
-						final long[] keyGroupRangeOffsets = new long[keyGroupRange.getNumberOfKeyGroups()];
-
-						for (int keyGroupPos = 0; keyGroupPos < keyGroupRange.getNumberOfKeyGroups(); ++keyGroupPos) {
-							int keyGroupId = keyGroupRange.getKeyGroupId(keyGroupPos);
-							keyGroupRangeOffsets[keyGroupPos] = localStream.getPos();
-							outView.writeInt(keyGroupId);
-
-							for (Map.Entry<StateUID, StateSnapshot> stateSnapshot :
-								cowStateStableSnapshots.entrySet()) {
-								StateSnapshot.StateKeyGroupWriter partitionedSnapshot =
-
-									stateSnapshot.getValue().getKeyGroupWriter();
-								try (
-									OutputStream kgCompressionOut =
-										keyGroupCompressionDecorator.decorateWithCompression(localStream)) {
-									DataOutputViewStreamWrapper kgCompressionView =
-										new DataOutputViewStreamWrapper(kgCompressionOut);
-									kgCompressionView.writeShort(stateNamesToId.get(stateSnapshot.getKey()));
-									partitionedSnapshot.writeStateInKeyGroup(kgCompressionView, keyGroupId);
-								} // this will just close the outer compression stream
-							}
-						}
-
-						if (snapshotCloseableRegistry.unregisterCloseable(streamWithResultProvider)) {
-							KeyGroupRangeOffsets kgOffs = new KeyGroupRangeOffsets(keyGroupRange, keyGroupRangeOffsets);
-							SnapshotResult<StreamStateHandle> result =
-								streamWithResultProvider.closeAndFinalizeCheckpointStreamResult();
-							return CheckpointStreamWithResultProvider.toKeyedStateHandleSnapshotResult(result, kgOffs);
-						} else {
-							throw new IOException("Stream already unregistered.");
-						}
-					}
-
-					@Override
-					protected void cleanupProvidedResources() {
-						for (StateSnapshot tableSnapshot : cowStateStableSnapshots.values()) {
-							tableSnapshot.release();
-						}
-					}
-
-					@Override
-					protected void logAsyncSnapshotComplete(long startTime) {
-						if (snapshotStrategySynchronicityTrait.isAsynchronous()) {
-							logAsyncCompleted(primaryStreamFactory, startTime);
-						}
-					}
-				};
-
-			final FutureTask<SnapshotResult<KeyedStateHandle>> task =
-				asyncSnapshotCallable.toAsyncSnapshotFutureTask(cancelStreamRegistry);
-			finalizeSnapshotBeforeReturnHook(task);
-
-			return task;
-		}
-
-		@Override
-		public void finalizeSnapshotBeforeReturnHook(Runnable runnable) {
-			snapshotStrategySynchronicityTrait.finalizeSnapshotBeforeReturnHook(runnable);
-		}
-
-		@Override
-		public boolean isAsynchronous() {
-			return snapshotStrategySynchronicityTrait.isAsynchronous();
-		}
-
-		@Override
-		public <N, V> StateTable<K, N, V> newStateTable(RegisteredKeyValueStateBackendMetaInfo<N, V> newMetaInfo) {
-			return snapshotStrategySynchronicityTrait.newStateTable(newMetaInfo);
-		}
-
-		private void processSnapshotMetaInfoForAllStates(
-			List<StateMetaInfoSnapshot> metaInfoSnapshots,
-			Map<StateUID, StateSnapshot> cowStateStableSnapshots,
-			Map<StateUID, Integer> stateNamesToId,
-			Map<String, ? extends StateSnapshotRestore> registeredStates,
-			StateMetaInfoSnapshot.BackendStateType stateType) {
-
-			for (Map.Entry<String, ? extends StateSnapshotRestore> kvState : registeredStates.entrySet()) {
-				final StateUID stateUid = StateUID.of(kvState.getKey(), stateType);
-				stateNamesToId.put(stateUid, stateNamesToId.size());
-				StateSnapshotRestore state = kvState.getValue();
-				if (null != state) {
-					final StateSnapshot stateSnapshot = state.stateSnapshot();
-					metaInfoSnapshots.add(stateSnapshot.getMetaInfoSnapshot());
-					cowStateStableSnapshots.put(stateUid, stateSnapshot);
-				}
-			}
-		}
-	}
-
 	private interface StateFactory {
 		<K, N, SV, S extends State, IS extends S> IS createState(
 			StateDescriptor<S, SV> stateDesc,
@@ -827,52 +387,4 @@ public class HeapKeyedStateBackend<K> extends AbstractKeyedStateBackend<K> {
 			TypeSerializer<K> keySerializer) throws Exception;
 	}
 
-	/**
-	 * Unique identifier for registered state in this backend.
-	 */
-	private static final class StateUID {
-
-		@Nonnull
-		private final String stateName;
-
-		@Nonnull
-		private final StateMetaInfoSnapshot.BackendStateType stateType;
-
-		StateUID(@Nonnull String stateName, @Nonnull StateMetaInfoSnapshot.BackendStateType stateType) {
-			this.stateName = stateName;
-			this.stateType = stateType;
-		}
-
-		@Nonnull
-		public String getStateName() {
-			return stateName;
-		}
-
-		@Nonnull
-		public StateMetaInfoSnapshot.BackendStateType getStateType() {
-			return stateType;
-		}
-
-		@Override
-		public boolean equals(Object o) {
-			if (this == o) {
-				return true;
-			}
-			if (o == null || getClass() != o.getClass()) {
-				return false;
-			}
-			StateUID uid = (StateUID) o;
-			return Objects.equals(getStateName(), uid.getStateName()) &&
-				getStateType() == uid.getStateType();
-		}
-
-		@Override
-		public int hashCode() {
-			return Objects.hash(getStateName(), getStateType());
-		}
-
-		public static StateUID of(@Nonnull String stateName, @Nonnull StateMetaInfoSnapshot.BackendStateType stateType) {
-			return new StateUID(stateName, stateType);
-		}
-	}
 }
