+++ flink-runtime/src/test/java/org/apache/flink/runtime/state/ttl/mock/MockKeyedStateBackend.java
@@ -30,7 +30,6 @@ import org.apache.flink.api.common.state.ValueStateDescriptor;
 import org.apache.flink.api.common.typeutils.TypeSerializer;
 import org.apache.flink.api.java.tuple.Tuple2;
 import org.apache.flink.core.fs.CloseableRegistry;
-import org.apache.flink.metrics.MetricGroup;
 import org.apache.flink.runtime.checkpoint.CheckpointOptions;
 import org.apache.flink.runtime.query.TaskKvStateRegistry;
 import org.apache.flink.runtime.state.AbstractKeyedStateBackend;
@@ -57,7 +56,6 @@ import javax.annotation.Nonnull;
 
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -87,9 +85,9 @@ public class MockKeyedStateBackend<K> extends AbstractKeyedStateBackend<K> {
 			Tuple2.of(FoldingStateDescriptor.class, (StateFactory) MockInternalFoldingState::createState)
 		).collect(Collectors.toMap(t -> t.f0, t -> t.f1));
 
-	private final Map<String, Map<K, Map<Object, Object>>> stateValues = new HashMap<>();
+	private final Map<String, Map<K, Map<Object, Object>>> stateValues;
 
-	private final Map<String, StateSnapshotTransformer<Object>> stateSnapshotFilters = new HashMap<>();
+	private final Map<String, StateSnapshotTransformer<Object>> stateSnapshotFilters;
 
 	MockKeyedStateBackend(
 		TaskKvStateRegistry kvStateRegistry,
@@ -99,10 +97,13 @@ public class MockKeyedStateBackend<K> extends AbstractKeyedStateBackend<K> {
 		KeyGroupRange keyGroupRange,
 		ExecutionConfig executionConfig,
 		TtlTimeProvider ttlTimeProvider,
-		MetricGroup operatorMetricGroup,
+		Map<String, Map<K, Map<Object, Object>>> stateValues,
+		Map<String, StateSnapshotTransformer<Object>> stateSnapshotFilters,
 		CloseableRegistry cancelStreamRegistry) {
 		super(kvStateRegistry, keySerializer, userCodeClassLoader,
 			numberOfKeyGroups, keyGroupRange, executionConfig, ttlTimeProvider, cancelStreamRegistry);
+		this.stateValues = stateValues;
+		this.stateSnapshotFilters = stateSnapshotFilters;
 	}
 
 	@Override
@@ -187,17 +188,10 @@ public class MockKeyedStateBackend<K> extends AbstractKeyedStateBackend<K> {
 	@SuppressWarnings("unchecked")
 	@Override
 	public void restore(Collection<KeyedStateHandle> state) {
-		stateValues.clear();
-		state = state == null ? Collections.emptyList() : state;
-		state.forEach(ksh -> stateValues.putAll(copy(((MockKeyedStateHandle<K>) ksh).snapshotStates)));
+		// all restore work done in builder and nothing to do here
 	}
 
-	private static <K> Map<String, Map<K, Map<Object, Object>>> copy(
-		Map<String, Map<K, Map<Object, Object>>> stateValues) {
-		return copy(stateValues, Collections.emptyMap());
-	}
-
-	private static <K> Map<String, Map<K, Map<Object, Object>>> copy(
+	static <K> Map<String, Map<K, Map<Object, Object>>> copy(
 		Map<String, Map<K, Map<Object, Object>>> stateValues, Map<String, StateSnapshotTransformer<Object>> stateSnapshotFilters) {
 		Map<String, Map<K, Map<Object, Object>>> snapshotStates = new HashMap<>();
 		for (String stateName : stateValues.keySet()) {
@@ -244,7 +238,7 @@ public class MockKeyedStateBackend<K> extends AbstractKeyedStateBackend<K> {
 			0);
 	}
 
-	private static class MockKeyedStateHandle<K> implements KeyedStateHandle {
+	static class MockKeyedStateHandle<K> implements KeyedStateHandle {
 		private static final long serialVersionUID = 1L;
 
 		final Map<String, Map<K, Map<Object, Object>>> snapshotStates;
